pyproject.toml
[project]
name = "wonderstreet-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.5",
    "fastapi>=0.121.2",
    "google-auth>=2.43.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pyairtable>=3.3.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.2",
    "pytest>=9.0.1",
]
